requirements/requirements.txt
oauth2client
datetime
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-telegram-bot
pysocks
